requirements.txt
appnope==0.1.2
APScheduler==3.8.1
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
black==21.9b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.9
click==8.0.3
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
Flask==2.0.2
idna==3.3
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
regex==2021.10.21
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sympy==1.8
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tomli==1.2.1
tornado==6.1
traitlets==5.1.0
tweepy==4.4.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
